+++ starkinfra/issuinginvoice/__issuinginvoice.py
@@ -118,3 +118,4 @@ def page(cursor=None, limit=None, after=None, before=None, status=None, tags=Non
         limit=limit,
         user=user,
     )
+