+++ tests/sdk/testIssuingTokenActivation.py
@@ -1,0 +1,35 @@
+import starkinfra
+from json import dumps, loads
+from unittest import TestCase, main
+from tests.utils.user import exampleProject
+from starkcore.error import InvalidSignatureError
+
+
+starkinfra.user = exampleProject
+
+
+class TestIssuingTokenParseRight(TestCase):
+    content = "{\"activationMethod\": {\"type\": \"text\", \"value\": \"** *****-5678\"}, \"tokenId\": \"5585821789122165\", \"tags\": [\"token\", \"user/1234\"], \"cardId\": \"5189831499972623\"}"
+    valid_signature = "MEUCIAxn0FmsPWI4r3Y7Nq8xFNQHYZgo0QAGDQ4/7CajKoVuAiEA09kXWrPMhsw4JbgC3pmNccCWr+hidfop/KsSNqza0yE="
+    invalid_signature = "MEUCIQDOpo1j+V40DNZK2URL2786UQK/8mDXon9ayEd8U0/l7AIgYXtIZJBTs8zCRR3vmted6Ehz/qfw1GRut/eYyvf1yOk="
+
+    def test_success(self):
+        event = starkinfra.issuingtokenactivation.parse(
+            content=self.content,
+            signature=self.valid_signature
+        )
+        print(event)
+
+    def test_normalized_success(self):
+        event = starkinfra.issuingtoken.parse(
+            content=dumps(loads(self.content), sort_keys=False, indent=4),
+            signature=self.valid_signature
+        )
+        print(event)
+
+    def test_invalid_signature(self):
+        with self.assertRaises(InvalidSignatureError):
+            starkinfra.issuingtoken.parse(
+                content=self.content,
+                signature=self.invalid_signature,
+            )